ChatDev/requirements.txt
arxiv==2.1.3
beautifulsoup4==4.13.5
chromadb==0.5.7
diskcache==5.6.3
easydict==1.10
easyocr==1.7.1
joblib
pyarrow
mammoth==1.8.0
markdownify==0.13.1
networkx==3.2.1
numpy==2.3.2
openai==1.107.0
openai_whisper
opencv_python_headless==4.10.0.84
pandas==2.3.2
pathvalidate==3.2.0
pdfminer.six==20240706
Pillow==11.3.0
puremagic==1.27
pydub==0.25.1
python_pptx==1.0.2
pytube==15.0.0
pyvis==0.3.2
PyYAML==6.0
Requests==2.32.5
seaborn==0.13.2
SpeechRecognition==3.10.4
sympy==1.12
tenacity==9.0.0
torch==2.4.0
tqdm==4.66.5
transformers==4.43.4
whisper==1.1.10
youtube_transcript_api==0.6.2flask
flask-cors
